Rename InlineConditions component file and adjust imports
